+++ age-core/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to Rust's notion of
 to 1.0.0 are beta releases.
 
 ## [Unreleased]
+### Added
+- `age_core::format::is_arbitrary_string`
+
+### Changed
+- `age::plugin::Connection::unidir_receive` now takes an additional argument to
+  enable handling an optional fourth command.
 
 ## [0.10.0] - 2024-02-04
 ### Added